requirements.txt
instagrapi>=1.18.0
openai>=1.14.0
python-dotenv>=1.0.0
Pillow>=9.0.0
requests>=2.28.0
